td/requirements.txt
python-socketio[client]
python-osc
